src/requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements/prod.txt requirements/in/prod.txt
#
attrs==21.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
autograd==1.3
    # via newsnlp
automat==20.2.0
    # via twisted
beautifulsoup4==4.10.0
    # via
    #   feedfinder2
    #   newspaper3k
blis==0.7.5
    # via
    #   spacy
    #   thinc
catalogue==2.0.6
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.7
    # via requests
click==7.1.2
    # via
    #   nltk
    #   sacremoses
    #   typer
commonmark==0.9.1
    # via rich
constantly==15.1.0
    # via twisted
cryptography==35.0.0
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.1.0
    # via
    #   newspaper3k
    #   parsel
    #   scrapy
cycler==0.10.0
    # via matplotlib
cymem==2.0.6
    # via
    #   preshed
    #   spacy
    #   thinc
cython==0.29.23
    # via newsnlp
daily-query @ git+https://github.com/techoutlooks/daily-query.git
    # via scrapy-newsutils
dnspython==2.2.1
    # via -r requirements/in/base.txt
feedfinder2==0.0.4
    # via newspaper3k
feedparser==6.0.8
    # via newspaper3k
filelock==3.3.1
    # via
    #   huggingface-hub
    #   tldextract
    #   transformers
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.0.0/fr_core_news_sm-3.0.0.tar.gz
    # via newsnlp
future==0.18.2
    # via autograd
h2==3.2.0
    # via
    #   scrapy
    #   twisted
hpack==3.0.0
    # via h2
huggingface-hub==0.11.1
    # via transformers
hyperframe==5.2.0
    # via h2
hyperlink==21.0.0
    # via twisted
idna==3.3
    # via
    #   hyperlink
    #   requests
    #   tldextract
image-quality==1.2.7
    # via
    #   -r requirements/in/base.txt
    #   scrapy-newsutils
imageio==2.9.0
    # via scikit-image
incremental==21.3.0
    # via twisted
itemadapter==0.4.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.0.4
    # via scrapy
jieba3k==0.35.1
    # via newspaper3k
jinja2==3.0.2
    # via spacy
jmespath==0.10.0
    # via itemloaders
joblib==1.1.0
    # via
    #   nltk
    #   sacremoses
kiwisolver==1.3.2
    # via matplotlib
libsvm==3.23.0.4
    # via image-quality
lxml==4.6.3
    # via
    #   newspaper3k
    #   parsel
    #   scrapy
markupsafe==2.0.1
    # via jinja2
matplotlib==3.4.3
    # via scikit-image
murmurhash==1.0.6
    # via
    #   preshed
    #   spacy
    #   thinc
networkx==2.6.3
    # via scikit-image
newsnlp @ git+https://github.com/techoutlooks/newsnlp.git
    # via -r requirements/in/prod.txt
newspaper3k==0.2.8
    # via
    #   -r requirements/in/base.txt
    #   scrapy-newsutils
nltk==3.6.2
    # via
    #   -r requirements/in/base.txt
    #   newspaper3k
    #   scrapy-newsutils
numpy==1.21.3
    # via
    #   autograd
    #   blis
    #   image-quality
    #   imageio
    #   matplotlib
    #   pywavelets
    #   scikit-image
    #   scipy
    #   spacy
    #   thinc
    #   tifffile
    #   transformers
ordered-set==4.1.0
    # via
    #   -r requirements/in/base.txt
    #   daily-query
    #   scrapy-newsutils
packaging==21.0
    # via
    #   huggingface-hub
    #   spacy
    #   transformers
parsel==1.6.0
    # via
    #   itemloaders
    #   scrapy
pathy==0.6.0
    # via spacy
pillow==8.2.0
    # via
    #   -r requirements/in/base.txt
    #   image-quality
    #   imageio
    #   matplotlib
    #   newspaper3k
    #   scikit-image
    #   scrapy-newsutils
preshed==3.0.6
    # via
    #   spacy
    #   thinc
priority==1.3.0
    # via twisted
protego==0.1.16
    # via scrapy
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycountry==20.7.3
    # via
    #   -r requirements/in/base.txt
    #   scrapy-newsutils
pycparser==2.20
    # via cffi
pydantic==1.7.4
    # via
    #   spacy
    #   thinc
pydispatcher==2.0.5
    # via scrapy
pygments==2.11.2
    # via rich
pymongo==3.11.4
    # via
    #   -r requirements/in/base.txt
    #   daily-query
    #   scrapy-newsutils
pyopenssl==21.0.0
    # via scrapy
pyparsing==3.0.0
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   newspaper3k
pywavelets==1.1.1
    # via scikit-image
pyyaml==6.0
    # via
    #   huggingface-hub
    #   newspaper3k
    #   transformers
queuelib==1.6.2
    # via scrapy
regex==2021.10.23
    # via
    #   nltk
    #   sacremoses
    #   transformers
requests==2.26.0
    # via
    #   feedfinder2
    #   huggingface-hub
    #   newspaper3k
    #   requests-file
    #   spacy
    #   tldextract
    #   transformers
requests-file==1.5.1
    # via tldextract
rich==12.0.1
    # via
    #   -r requirements/in/base.txt
    #   scrapy-newsutils
sacremoses==0.0.53
    # via newsnlp
schedule==1.1.0
    # via
    #   -r requirements/in/base.txt
    #   scrapy-newsutils
scikit-image==0.18.3
    # via image-quality
scipy==1.7.1
    # via
    #   image-quality
    #   scikit-image
scrapy==2.5.0
    # via
    #   -r requirements/in/base.txt
    #   scrapy-newsutils
scrapy-newsutils @ git+https://github.com/techoutlooks/scrapy-newsutils.git
    # via -r requirements/in/prod.txt
sentencepiece==0.1.97
    # via newsnlp
service-identity==21.1.0
    # via scrapy
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   automat
    #   feedfinder2
    #   parsel
    #   protego
    #   pyopenssl
    #   python-dateutil
    #   requests-file
    #   sacremoses
    #   service-identity
    #   w3lib
smart-open==5.2.1
    # via pathy
soupsieve==2.2.1
    # via beautifulsoup4
spacy==3.0.6
    # via
    #   fr-core-news-sm
    #   newsnlp
spacy-legacy==3.0.8
    # via spacy
srsly==2.4.2
    # via
    #   spacy
    #   thinc
thinc==8.0.11
    # via spacy
tifffile==2021.10.12
    # via scikit-image
tinysegmenter==0.3
    # via newspaper3k
tldextract==3.1.2
    # via newspaper3k
tokenizers==0.13.2
    # via transformers
torch==1.13.1
    # via transformers
tqdm==4.62.3
    # via
    #   huggingface-hub
    #   nltk
    #   sacremoses
    #   spacy
    #   transformers
transformers[torch]==4.25.1
    # via newsnlp
twisted[http2]==21.7.0
    # via scrapy
typer==0.3.2
    # via
    #   pathy
    #   spacy
typing-extensions==3.10.0.2
    # via
    #   huggingface-hub
    #   torch
    #   twisted
urllib3==1.26.7
    # via requests
w3lib==1.22.0
    # via
    #   itemloaders
    #   parsel
    #   scrapy
wasabi==0.8.2
    # via
    #   spacy
    #   thinc
zope-interface==5.4.0
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
